LayoutTests/imported/w3c/web-platform-tests/css/css-position/overlay/overlay-user-agent-rules-expected.txt


FAIL HTML elements should have overlay:none !important from the user-agent. assert_equals: expected (string) "none" but got (undefined) undefined
FAIL SVG elements should have overlay:none !important from the user-agent. assert_equals: expected (string) "none" but got (undefined) undefined
FAIL Null namespace elements should have overlay:none !important from the user-agent. assert_equals: expected (string) "none" but got (undefined) undefined
FAIL Arbitrary namespace elements should have overlay:none !important from the user-agent. assert_equals: expected (string) "none" but got (undefined) undefined
